Use shorts instead of ints where appropriate; hook up creating and enqueuing animations from the server POST handler

+++ led-arch/lib/Fx_Animations.h
@@ -7,13 +7,13 @@
 // -----------------------------------------
 // Declare Animation functions
 
-void colorBlink(float progress, RGBColor* pixels, int pixelCount, Fx_AnimationParams *params);
+void colorBlink(float progress, RGBColor* pixels, unsigned short pixelCount, Fx_AnimationParams *params);
 
-void allOff(float progress, RGBColor* pixels, int pixelCount, Fx_AnimationParams *params);
+void allOff(float progress, RGBColor* pixels, unsigned short pixelCount, Fx_AnimationParams *params);
 
-void endWithColor(float progress, RGBColor* pixels, int pixelCount, Fx_AnimationParams *params);
+void endWithColor(float progress, RGBColor* pixels, unsigned short pixelCount, Fx_AnimationParams *params);
 
-void bounce(float progress, RGBColor* pixels, int pixelCount, Fx_AnimationParams *params);
+void bounce(float progress, RGBColor* pixels, unsigned short pixelCount, Fx_AnimationParams *params);
 
 // ------------------------
 // Define some pre-defined colors
